Still works after minor refactoring.
