--- src/main/kotlin/com/example/models/YandexResponse.kt
@@ -1,5 +1,0 @@
-package com.example.models
-
-data class YandexResponse (
-    val result: String
-)
