+++ src/main/kotlin/com/example/plugins/Clients.kt
@@ -1,0 +1,16 @@
+package com.example.plugins
+
+import io.ktor.client.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.client.plugins.logging.*
+import io.ktor.serialization.jackson.*
+
+fun configureClient(): HttpClient = HttpClient(CIO) {
+    install(Logging) {
+        level = LogLevel.ALL
+    }
+    install(ContentNegotiation) {
+        jackson()
+    }
+}
