+++ src/main/kotlin/com/example/plugins/Routing.kt
@@ -1,16 +1,17 @@
 package com.example.plugins
 
+import com.example.adapters.AdapterType
 import com.example.routes.*
 import io.ktor.client.*
 import io.ktor.server.application.*
 import io.ktor.server.routing.*
 
-fun Application.configureRouting(client: HttpClient) {
+fun Application.configureRouting(client: HttpClient, aiAdapterType: AdapterType) {
     routing {
         listAllRequests()
         getRequestById()
         createRequest()
         deleteRequest()
-        proxyRequest(client)
+        proxyRequest(client, aiAdapterType)
     }
 }
