+++ src/main/resources/application.yaml
@@ -2,5 +2,6 @@ ktor:
     application:
         modules:
             - com.example.ApplicationKt.module
+        ai_adapter: Yandex
     deployment:
         port: 12222
